move go files to non top level location
The intent of this change is to hid the go code from the top level of
the repo, so that the bindings can be generated in this repo, and still
be included in external projects if this repo is used as a submodule.
It is also a valid use case to build the bindings, and then copy them to
another project to be built as if they are a part of that projects go
module. This eases dependency conflicts!