Sync release tee caff node (#23)
* Move gen.go to top level of project to support building from nitro-repo

It seems that go doesn't like the go.mod file and the file with main()
in it to be in separate locations when being imported in another
project. Additonally, it seems that the override to geth in the
nitro-repo is replacing the import to geth here. Fun times!

* move go files to non top level location

The intent of this change is to hid the go code from the top level of
the repo, so that the bindings can be generated in this repo, and still
be included in external projects if this repo is used as a submodule.
It is also a valid use case to build the bindings, and then copy them to
another project to be built as if they are a part of that projects go
module. This eases dependency conflicts!

* Fix location of generated go binding file.