Add MultiSigTransfer.s.sol
This commit adds a forge script that will initiate ownership transfer of
all EspressoTEEVerifier contracts for a network to a new address
provided in an environment variable.
The core logic for this script has been tested in a solidity test to
ensure that it will enable valid ownership transfer for all nested
contracts.