Zs/multi sig ownership transfer (#31)
* Add test for logic of MultiSigTransfer.s.sol

* Add MultiSigTransfer.s.sol

This commit adds a forge script that will initiate ownership transfer of
all EspressoTEEVerifier contracts for a network to a new address
provided in an environment variable.
The core logic for this script has been tested in a solidity test to
ensure that it will enable valid ownership transfer for all nested
contracts.

* run `forge fmt`

* remove extraneous comments

* run `forge fmt`

* Remove unused imports, and improve logging

* Update comment

* Respond to comments and add better assertions in script.

This comment applies many formatting fixes, responds to most comments,
and adds better assertions about the pending state of ownership transfer
to the `MultiSigTransfer.s.sol` script.