Propose multi-sig transaction to accept ownership transfer of TEEVerifier contracts. (#32)
* Add safe-utils import

This adds the safe-utils dependency to begin adding support for
proposing multi-sig transactions.

* slightly refactor tests

* Add initial support for proposing multi-sig transactions to the safe web API

* Update MultiSigTransfer script to support proposing transactions

Currently, this only supports ETH, ETH Sepolia, and ARB One. We are
hoping to support Arb Sepolia in the future via extending safe-utils to
utilize OnChainDen

* Update remappings.txt to fix path forge was searching for safe-transaction-service imports

While compiling the safe-utils dependency, forge attempted to look for
safe-transaction-service imports in the wrong path. This commit fixes
that issue.

* Update README.md

This commit updates README.md to provide insight into, and instructions
for running the MultiSigTransfer script to transfer an
EspressoTEEVerifier contract from an existing EOA, and propose a
transaction on a destination Safe multi-sig wallet via the
safe-transaction-service API

* run forge fmt

* fix remappings for CI (and others)

* fix remappings for real this time

* respond to nits

* Update scripts/MultiSigTransfer.s.sol

Co-authored-by: Alysia Tech <hi@alysia.tech>

---------

Co-authored-by: Alysia Tech <hi@alysia.tech>